Workshop Week 3 Task 1,2 and 3 added

+++ routes/campsiteRouter.js
@@ -17,27 +17,39 @@ campsiteRouter.route('/')
   .catch(err => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
-  Campsite.create(req.body)
-  .then(campsite => {
-    console.log('Campsite created ', campsite);
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.json(campsite);
-  })
-  .catch(err => next(err));
+  if(req.user.admin) {  
+    Campsite.create(req.body)
+    .then(campsite => {
+      console.log('Campsite created ', campsite);
+      res.statusCode = 200;
+      res.setHeader('Content-Type', 'application/json');
+      res.json(campsite);
+    })
+    .catch(err => next(err));
+  } else {
+    err = new Error('You are not authorized to perform this operation');
+    err.status = 403;
+    return next(err);
+  }
 })
 .put(authenticate.verifyUser, (req, res) => {
   res.statusCode = 403;
   res.end('PUT operation not supported on /campsites');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-   Campsite.deleteMany()
+  if(req.user.admin) {
+    Campsite.deleteMany()
    .then(response => {
      res.statusCode = 200;
      res.setHeader('Content-Type', 'application/json');
      res.json(response);
-   })
+    })
    .catch(err => next(err));
+  } else {
+    err = new Error('You are not authorized to perform this operation');
+    err.status = 403;
+    return next(err);
+  }
 });
 
 campsiteRouter.route('/:campsiteId')
@@ -56,24 +68,36 @@ campsiteRouter.route('/:campsiteId')
   res.end(`POST operation not supported on /campsites/${req.params.campsiteId}`);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+  if(req.user.admin) {
   Campsite.findByIdAndUpdate(req.params.campsiteId, {
     $set: req.body
-}, { new: true })
-.then(campsite => {
+    }, { new: true })
+   .then(campsite => {
     res.statusCode = 200;
     res.setHeader('Content-Type', 'application/json');
     res.json(campsite);
-})
-.catch(err => next(err));
+  })
+  .catch(err => next(err));
+  } else {
+    err = new Error('You are not authorized to perform this operation');
+    err.status = 403;
+    return next(err);
+  }
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-  Campsite.findByIdAndDelete(req.params.campsiteId)
+  if(req.user.admin) {
+    Campsite.findByIdAndDelete(req.params.campsiteId)
     .then(response => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
         res.json(response);
     })
     .catch(err => next(err));
+  } else {
+    err = new Error('You are not authorized to perform this operation');
+    err.status = 403;
+    return next(err);
+  }
 });
 
 campsiteRouter.route('/:campsiteId/comments')
@@ -119,6 +143,7 @@ campsiteRouter.route('/:campsiteId/comments')
     res.end(`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`);
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+  if(req.user.admin) {
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
         if (campsite) {
@@ -139,6 +164,11 @@ campsiteRouter.route('/:campsiteId/comments')
         }
     })
     .catch(err => next(err));
+  } else {
+    err = new Error('You are not authorized to perform this operation');
+    err.status = 403;
+    return next(err);
+  }
 });
 
 campsiteRouter.route('/:campsiteId/comments/:commentId')
